requirements.txt
asgiref==3.9.1
bleach==4.1.0
crispy-bootstrap5==2025.6
Django==5.2.4
django-appconf==1.1.0
django-crispy-forms==2.4
django-imagekit==5.0.0
django-summernote==0.8.20.0
pilkit==3.0
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
webencodings==0.5.1
